16/02

+++ ApiProdutos/src/main/java/br/com/bling/ApiProdutos/controllers/ProdutoController.java
@@ -1,7 +1,11 @@
 package br.com.bling.ApiProdutos.controllers;
 
+import br.com.bling.ApiProdutos.exceptions.ProdutoCadastroException;
+import br.com.bling.ApiProdutos.exceptions.ProdutoNaoEncontradoException;
+import br.com.bling.ApiProdutos.exceptions.ProdutoNaoEncontradoParaExclusaoException;
 import br.com.bling.ApiProdutos.models.Produto;
 import br.com.bling.ApiProdutos.service.ProdutoService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
@@ -9,6 +13,8 @@ import org.springframework.web.bind.annotation.*;
 
 import br.com.bling.ApiProdutos.models.Resposta;
 import io.swagger.annotations.Api;
+
+import java.util.List;
 
 
 @RestController
@@ -42,10 +48,15 @@ public class ProdutoController {
     public Resposta getProductByCode(@PathVariable String codigo) {
         Resposta response = produtoService.getProductByCode(codigo);
 
+        if (response == null || response.getRetorno() == null) {
+            throw new ProdutoNaoEncontradoException(codigo);
+        }
+
         System.out.println(response);
 
         return response;
     }
+
 
     /**
      * GET "PRODUTO UTILIZANDO O CÓDIGO E NOME DO FORNECEDOR.
@@ -55,10 +66,15 @@ public class ProdutoController {
     public Resposta getProductByCodeSupplier(@PathVariable String codigo, String nomeFornecedor) {
         Resposta response = produtoService.getProductByCodeSupplier(codigo, nomeFornecedor);
 
+        if (response == null || response.getRetorno() == null) {
+            throw new ProdutoNaoEncontradoException(codigo);
+        }
+
         System.out.println(response);
 
         return response;
     }
+
 
     /**
      * DELETE PROUTO PELO CÓDIGO (SKU).
@@ -66,23 +82,34 @@ public class ProdutoController {
     @DeleteMapping("/produto/{codigo}")
     @ApiOperation(value = "Deletar um produto pelo código")
     public void deleteProductByCode(@PathVariable String codigo) {
+        Resposta response = produtoService.getProductByCode(codigo);
+
+        if (response == null || response.getRetorno() == null) {
+            throw new ProdutoNaoEncontradoParaExclusaoException(codigo);
+        }
+
         produtoService.deleteProductByCode(codigo);
 
         System.out.println("Codigo deletado = " + codigo);
     }
+
 
     /**
      * POST "CADASTRAR UM NOVO PRODUTO" UTILIZANDO XML.
      */
     @PostMapping(path = "/cadastrarproduto", consumes = MediaType.APPLICATION_XML_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation(value = "Cadastrar um novo produto")
+    @ResponseStatus(HttpStatus.CREATED)
     public String createProduct(@RequestBody String xml) {
-        String request = produtoService.createProduct(xml);
-
-        System.out.println(request);
-
-        return request;
+        try {
+            String request = produtoService.createProduct(xml);
+            System.out.println(request);
+            return request;
+        } catch (Exception e) {
+            throw new ProdutoCadastroException("Erro ao cadastrar produto: " + e.getMessage());
+        }
     }
+
 
     /* - EXEMPLO DO VIDEO COM O DE BAIXO
